Organize classes into folders
