+++ SocialEventManager/src/SocialEventManager.BLL/GlobalSuppressions.cs
@@ -1,4 +1,4 @@
 using System.Diagnostics.CodeAnalysis;
 
-[assembly: SuppressMessage("Usage", "CA1816:Dispose methods should call SuppressFinalize", Justification = "<Pending>", Scope = "member", Target = "~M:SocialEventManager.BLL.Services.CustomUsersStore.Dispose")]
-[assembly: SuppressMessage("Usage", "CA1816:Dispose methods should call SuppressFinalize", Justification = "<Pending>", Scope = "member", Target = "~M:SocialEventManager.BLL.Services.CustomRolesStore.Dispose")]
+[assembly: SuppressMessage("Usage", "CA1816:Dispose methods should call SuppressFinalize", Justification = "<Pending>", Scope = "member", Target = "~M:SocialEventManager.BLL.Services.Identity.CustomRolesStore.Dispose")]
+[assembly: SuppressMessage("Usage", "CA1816:Dispose methods should call SuppressFinalize", Justification = "<Pending>", Scope = "member", Target = "~M:SocialEventManager.BLL.Services.Identity.CustomUsersStore.Dispose")]
