+++ SocialEventManager/src/SocialEventManager.BLL/MappingProfiles/AccountProfile.cs
@@ -1,6 +1,6 @@
 using System;
 using AutoMapper;
-using SocialEventManager.BLL.Models;
+using SocialEventManager.BLL.Models.Accounts;
 using SocialEventManager.BLL.Models.Identity;
 using SocialEventManager.DLL.Entities;
