+++ SocialEventManager/src/SocialEventManager.BLL/MappingProfiles/UserProfile.cs
@@ -1,6 +1,7 @@
 using AutoMapper;
 using SocialEventManager.BLL.Models;
 using SocialEventManager.BLL.Models.Identity;
+using SocialEventManager.BLL.Models.Users;
 using SocialEventManager.DLL.Entities;
 
 namespace SocialEventManager.BLL.MappingProfiles
