+++ SocialEventManager/src/SocialEventManager.BLL/Models/Accounts/AccountDto.cs
@@ -1,4 +1,4 @@
-namespace SocialEventManager.BLL.Models
+namespace SocialEventManager.BLL.Models.Accounts
 {
     public class AccountDto : BaseAccountDto
     {
