+++ SocialEventManager/src/SocialEventManager.BLL/Models/Accounts/BaseAccountDto.cs
@@ -4,7 +4,7 @@ using Dapper.Contrib.Extensions;
 using SocialEventManager.Shared.Constants;
 using SocialEventManager.Shared.Utilities.Attributes;
 
-namespace SocialEventManager.BLL.Models
+namespace SocialEventManager.BLL.Models.Accounts
 {
     public class BaseAccountDto
     {
