+++ SocialEventManager/src/SocialEventManager.BLL/Models/Roles/RoleForCreationDto.cs
@@ -1,4 +1,4 @@
-namespace SocialEventManager.BLL.Models
+namespace SocialEventManager.BLL.Models.Roles
 {
     public class RoleForCreationDto : BaseRoleDto
     {
