+++ SocialEventManager/src/SocialEventManager.BLL/Models/Roles/RoleForUpdateDto.cs
@@ -1,4 +1,4 @@
-namespace SocialEventManager.BLL.Models
+namespace SocialEventManager.BLL.Models.Roles
 {
     public class RoleForUpdateDto : BaseRoleDto
     {
