+++ SocialEventManager/src/SocialEventManager.BLL/Models/Users/LoginModel.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace SocialEventManager.BLL.Models
+namespace SocialEventManager.BLL.Models.Users
 {
     public class LoginModel
     {
