+++ SocialEventManager/src/SocialEventManager.BLL/Services/Users/IUserRolesService.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using SocialEventManager.BLL.Models;
+using SocialEventManager.BLL.Models.Users;
 
-namespace SocialEventManager.BLL.Services
+namespace SocialEventManager.BLL.Services.Users
 {
     public interface IUserRolesService
     {
