+++ SocialEventManager/src/SocialEventManager.DLL/Repositories/Accounts/AccountsRepository.cs
@@ -1,7 +1,7 @@
 using SocialEventManager.DLL.Entities;
 using SocialEventManager.DLL.Infrastructure;
 
-namespace SocialEventManager.DLL.Repositories
+namespace SocialEventManager.DLL.Repositories.Accounts
 {
     public class AccountsRepository : GenericRepository<Account>, IAccountsRepository
     {
