+++ SocialEventManager/src/SocialEventManager.DLL/Repositories/Accounts/IAccountsRepository.cs
@@ -1,7 +1,7 @@
 using SocialEventManager.DLL.Entities;
 using SocialEventManager.DLL.Infrastructure;
 
-namespace SocialEventManager.DLL.Repositories
+namespace SocialEventManager.DLL.Repositories.Accounts
 {
     public interface IAccountsRepository : IGenericRepository<Account>
     {
